requirements.txt
python-dotenv~=1.0.1
openai~=1.13.3
discord.py
tiktoken~=0.7.0
discord~=2.3.2
aiohttp~=3.7.4.post0